requirements.txt
appdirs==1.4.4
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.2
fonttools==4.38.0
greenlet==2.0.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
multitasking==0.0.11
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
tfinance==0.1.3
threadpoolctl==3.1.0
tqdm==4.64.1
urllib3==1.26.12
Werkzeug==2.2.2
yfinance==0.1.85
zipp==3.10.0
alembic==1.8.1
anyio==3.6.2
apache-airflow==2.4.3
apache-airflow-providers-common-sql==1.2.0
apache-airflow-providers-ftp==3.1.0
apache-airflow-providers-http==4.0.0
apache-airflow-providers-imap==3.0.0
apache-airflow-providers-sqlite==3.2.1
apispec==3.3.2
appdirs==1.4.4
argcomplete==2.0.0
attrs==22.1.0
Babel==2.11.0
beautifulsoup4==4.11.1
blinker==1.5
bs4==0.0.1
cachelib==0.9.0
cattrs==22.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.1
contourpy==1.0.6
cron-descriptor==1.2.31
croniter==1.3.7
cryptography==38.0.3
cycler==0.11.0
Deprecated==1.2.13
dill==0.3.6
dnspython==2.2.1
docutils==0.19
email-validator==1.3.0
exceptiongroup==1.0.2
Flask==2.2.2
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.1
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fonttools==4.38.0
graphviz==0.20.1
greenlet==2.0.1
gunicorn==20.1.0
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
kiwisolver==1.4.4
lazy-object-proxy==1.8.0
linkify-it-py==2.0.0
lockfile==0.12.2
lxml==4.9.1
Mako==1.2.3
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.6.2
mdit-py-plugins==0.3.1
mdurl==0.1.2
multitasking==0.0.11
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pathspec==0.9.0
pendulum==2.1.2
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
pluggy==1.0.0
prison==0.2.1
psutil==5.9.4
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-daemon==2.3.2
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==6.1.2
pytz==2022.6
pytzdata==2020.1
PyYAML==6.0
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
rich==12.6.0
scikit-learn==1.1.3
scipy==1.9.3
setproctitle==1.3.2
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.3
sqlparse==0.4.3
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tenacity==8.1.0
termcolor==2.1.0
text-unidecode==1.3
tfinance==0.1.3
threadpoolctl==3.1.0
tqdm==4.64.1
typing-extensions==4.4.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.12
Werkzeug==2.2.2
wrapt==1.14.1
WTForms==3.0.1
yfinance==0.1.85
zipp==3.10.0
